# fixes #13618 - cache expensive vmware api calls