--- a/app/models/compute_resources/foreman/model/vmware.rb
+++ b/app/models/compute_resources/foreman/model/vmware.rb
@@ -4,6 +4,7 @@ require 'foreman/exception'
 module Foreman::Model
   class Vmware < ComputeResource
     include ComputeResourceConsoleCommon
+    include ComputeResourceCaching
 
     validates :user, :password, :server, :datacenter, :presence => true
     before_create :update_public_key
@@ -53,41 +54,54 @@ module Foreman::Model
     end
 
     def datacenters
-      name_sort(client.datacenters.all)
+      cache.cache(:datacenters) do
+        name_sort(client.datacenters.all)
+      end
     end
 
     def cluster(cluster)
-      dc.clusters.get(cluster)
+      cache.cache(:"cluster-#{cluster}") do
+        dc.clusters.get(cluster)
+      end
     end
 
     def clusters
-      if dc.clusters.nil?
+      dc_clusters = dc.clusters
+      if dc_clusters.nil?
         Rails.logger.info "Datacenter #{dc.try(:name)} returned zero clusters"
         return []
       end
-      dc.clusters.map(&:full_path).sort
+      dc_clusters.map(&:full_path).sort
     end
 
     def datastores(opts = {})
       if opts[:storage_domain]
-        name_sort(dc.datastores.get(opts[:storage_domain]))
+        cache.cache(:"datastores-#{opts[:storage_domain]}") do
+          name_sort(dc.datastores.get(opts[:storage_domain]))
+        end
       else
-        name_sort(dc.datastores.all(:accessible => true))
+        cache.cache(:datastores) do
+          name_sort(dc.datastores.all(:accessible => true))
+        end
       end
     end
 
     def storage_pods(opts = {})
       if opts[:storage_pod]
-        begin
-          dc.storage_pods.get(opts[:storage_pod])
-        rescue RbVmomi::VIM::InvalidArgument
-          {} # Return an empty storage pod hash if vsphere does not support the feature
+        cache.cache(:"storage_pods-#{opts[:storage_pod]}") do
+          begin
+            dc.storage_pods.get(opts[:storage_pod])
+          rescue RbVmomi::VIM::InvalidArgument
+            {} # Return an empty storage pod hash if vsphere does not support the feature
+          end
         end
       else
-        begin
-          name_sort(dc.storage_pods.all())
-        rescue RbVmomi::VIM::InvalidArgument
-          [] # Return an empty set of storage pods if vsphere does not support the feature
+        cache.cache(:storage_pods) do
+          begin
+            name_sort(dc.storage_pods.all())
+          rescue RbVmomi::VIM::InvalidArgument
+            [] # Return an empty set of storage pods if vsphere does not support the feature
+          end
         end
       end
     end
@@ -97,20 +111,28 @@ module Foreman::Model
     end
 
     def folders
-      dc.vm_folders.sort_by{|f| [f.path, f.name]}
+      cache.cache(:folders) do
+        dc.vm_folders.sort_by{|f| [f.path, f.name]}
+      end
     end
 
     def networks(opts = {})
-      name_sort(dc.networks.all(:accessible => true))
+      cache.cache(:networks) do
+        name_sort(dc.networks.all(:accessible => true))
+      end
     end
 
     def resource_pools(opts = {})
       cluster = cluster(opts[:cluster_id])
-      name_sort(cluster.resource_pools.all(:accessible => true))
+      cache.cache(:resource_pools) do
+        name_sort(cluster.resource_pools.all(:accessible => true))
+      end
     end
 
     def available_clusters
-      name_sort(dc.clusters)
+      cache.cache(:clusters) do
+        name_sort(dc.clusters)
+      end
     end
 
     def available_folders