--- a/app/models/concerns/compute_resource_caching.rb
+++ b/app/models/concerns/compute_resource_caching.rb
@@ -1,0 +1,13 @@
+module ComputeResourceCaching
+  extend ActiveSupport::Concern
+
+  def refresh_cache
+    cache.refresh
+  end
+
+  private
+
+  def cache
+    @cache ||= ComputeResourceCache.new(self)
+  end
+end