--- a/app/views/api/v2/compute_resources/vmware.json.rabl
+++ b/app/views/api/v2/compute_resources/vmware.json.rabl
@@ -1,1 +1,1 @@
-attributes :user, :datacenter, :server, :set_console_password
+attributes :user, :datacenter, :server, :set_console_password, :caching_enabled