--- a/db/migrate/20170127094357_add_caching_enabled_to_compute_resource.rb
+++ b/db/migrate/20170127094357_add_caching_enabled_to_compute_resource.rb
@@ -1,0 +1,5 @@
+class AddCachingEnabledToComputeResource < ActiveRecord::Migration
+  def change
+    add_column :compute_resources, :caching_enabled, :boolean, default: true
+  end
+end