@ohadlevy 2017-01-29
I would like to see a refresh / reload button on the host edit page, i think this is where you would really need  up to date / latest info as you dont have the option (dc, cluster etc).