@timogoebel 2017-01-29
> I would like to see a refresh / reload button on the host edit page, i think this is where you would really need up to date / latest info as you dont have the option (dc, cluster etc).

@ohadlevy : Would it be ok for you to add this in a separate PR? I think, this will make coding and reviewing easier.