@timogoebel 2017-01-30
@tbrisker , @ohadlevy : Addressed all of your comments. Thanks a lot for these.

> As for the naming of the action, "reload" seems to be a bit misleading, as it isn't actually reloaded, rather "cleared" or "flushed" IIUC.

@tbrisker : Would `refresh cache` be fine for your? I had `clear cache` before, but @ohadlevy suggested to use something different.

I currently see these follow up tasks:

- [ ] Add a rake task to refresh the cache.
- [ ] Add a cache refresh button to the select boxed in the host form that flushes the cache and reloads the data.