@timogoebel 2017-01-30
@shlomizadok : Would you mind testing this for me? I think a VM show and create should suffice.