@timogoebel 2017-01-30
Caching is now disabled if the compute resource has not been saved yet, fixed the kwarg default and removed caching from `dc` method as method is quite fast.