@timogoebel 2017-01-30
Rebased and renamed `reload_cache` to `refresh_cache`.