@timogoebel 2017-01-31
Fixed an issue with caching of `dc.clusters`.

@TheKangaroo: Would you mind testing this PR and let us know if you see any issues? Thanks.