@TheKangaroo 2017-01-31
Hi @tbrisker, I tested it in our environment.
I couldn't find any issues and the runtime improvements in creating hosts is huge :)
I really like your PR @timogoebel