@timogoebel 2017-01-31
Created two follow-up tickets:

Cache-Warming rake task:
http://projects.theforeman.org/issues/18314

Refresh from host page:
http://projects.theforeman.org/issues/18315